Add files via upload
So, as far as I know, this setup is "working" (adjusting the potentiometer on the Particle is actually registered on the Arduino, and it shows changing values on the Arduino side).
For some reason, though, it's showing strings of 0's between each "real" value (1-~180), so it's trying to move the arm extremely fast between the desired "1-180" value, and 0. This causes the arm to just kind of sit and "shake" as it tries to move between two positions almost instantly.
I've tried using "PULSE_IN," instead of the set up uploaded here, but when I try to use pulse_in, it shows either just an ever growing line of "?"'s, or it just shows a solid, unchanging value (usually somewhere in the 1200 range, even though the particle pin is mapped to send 0-255) no matter how the potentiometer is adjusted.
If you've got a potentiometer, Particle, and another board (Teensy, Arduino, second Particle, etc), you should be able to get the sketches uploaded to them and atleast see the "0"'s problem I'm having, and also can try to use pulse_in and see its "?"'s or unchanging values.

Maybe obvious, but the "working particle side" .ino file goes into the Particle IDE, and the "running_arduino_side" goes into the Arduino IDE. 
To test the Particle side: With the Particle connected,  just have the Arduino IDE open next to the Particle IDE. Then in the Arduino IDE, go to Tools, change the port to the Particle's port ("XXXXXusb modemXXXXXX" for me), and click the serial monitor button in Arduino and it'll show the Particle's serial output. 
Slack or text (501-282-NineSevenTwoEight) me if you have any questions, updates, suggestions, or want me to try running something you made. 
Thanks! 
-Alex